Address remaining CR comments for the ESP32 Azure IoT Kit sample
