+++ demos/sample_azure_iot_pnp/sample_azure_iot_pnp.c
@@ -18,540 +18,598 @@
 #include "azure_iot_json_reader.h"
 #include "azure_iot_json_writer.h"
 
-#include "sample_azure_iot_pnp.h"
+/* Exponential backoff retry include. */
+#include "backoff_algorithm.h"
 
+/* Transport interface implementation include header for TLS. */
+#include "transport_tls_socket.h"
+
+/* Crypto helper header. */
+#include "crypto.h"
+
+/* Demo Specific configs. */
+#include "demo_config.h"
+
+/* Data Interface Definition */
+#include "sample_azure_iot_pnp.h"
+/*-----------------------------------------------------------*/
+
+/* Compile time error for undefined configs. */
+#if !defined( democonfigHOSTNAME ) && !defined( democonfigENABLE_DPS_SAMPLE )
+    #error "Define the config democonfigHOSTNAME by following the instructions in file demo_config.h."
+#endif
+
+#if !defined( democonfigENDPOINT ) && defined( democonfigENABLE_DPS_SAMPLE )
+    #error "Define the config dps endpoint by following the instructions in file demo_config.h."
+#endif
+
+#ifndef democonfigROOT_CA_PEM
+    #error "Please define Root CA certificate of the IoT Hub(democonfigROOT_CA_PEM) in demo_config.h."
+#endif
+
+#if defined( democonfigDEVICE_SYMMETRIC_KEY ) && defined( democonfigCLIENT_CERTIFICATE_PEM )
+    #error "Please define only one auth democonfigDEVICE_SYMMETRIC_KEY or democonfigCLIENT_CERTIFICATE_PEM in demo_config.h."
+#endif
+
+#if !defined( democonfigDEVICE_SYMMETRIC_KEY ) && !defined( democonfigCLIENT_CERTIFICATE_PEM )
+    #error "Please define one auth democonfigDEVICE_SYMMETRIC_KEY or democonfigCLIENT_CERTIFICATE_PEM in demo_config.h."
+#endif
 /*-----------------------------------------------------------*/
 
 /**
- * @brief Command values
+ * @brief The maximum number of retries for network operation with server.
  */
-#define sampleazureiotCOMMAND_MAX_MIN_REPORT                  "getMaxMinReport"
-#define sampleazureiotCOMMAND_MAX_TEMP                        "maxTemp"
-#define sampleazureiotCOMMAND_MIN_TEMP                        "minTemp"
-#define sampleazureiotCOMMAND_TEMP_VERSION                    "avgTemp"
-#define sampleazureiotCOMMAND_START_TIME                      "startTime"
-#define sampleazureiotCOMMAND_END_TIME                        "endTime"
-#define sampleazureiotCOMMAND_EMPTY_PAYLOAD                   "{}"
-#define sampleazureiotCOMMAND_FAKE_END_TIME                   "2023-01-10T10:00:00Z"
+#define sampleazureiotRETRY_MAX_ATTEMPTS                      ( 5U )
 
 /**
- * @brief Device values
+ * @brief The maximum back-off delay (in milliseconds) for retrying failed operation
+ *  with server.
  */
-#define sampleazureiotDEFAULT_START_TEMP_COUNT                1
-#define sampleazureiotDEFAULT_START_TEMP_CELSIUS              22.0
-#define sampleazureiotDOUBLE_DECIMAL_PLACE_DIGITS             2
+#define sampleazureiotRETRY_MAX_BACKOFF_DELAY_MS              ( 5000U )
 
 /**
- * @brief Property Values
+ * @brief The base back-off delay (in milliseconds) to use for network operation retry
+ * attempts.
  */
-#define sampleazureiotPROPERTY_STATUS_SUCCESS                 200
-#define sampleazureiotPROPERTY_SUCCESS                        "success"
-#define sampleazureiotPROPERTY_TARGET_TEMPERATURE_TEXT        "targetTemperature"
-#define sampleazureiotPROPERTY_MAX_TEMPERATURE_TEXT           "maxTempSinceLastReboot"
+#define sampleazureiotRETRY_BACKOFF_BASE_MS                   ( 500U )
 
 /**
- * @brief Telemetry values
+ * @brief Timeout for receiving CONNACK packet in milliseconds.
  */
-#define sampleazureiotTELEMETRY_NAME                          "temperature"
+#define sampleazureiotCONNACK_RECV_TIMEOUT_MS                 ( 10 * 1000U )
 
 /**
- *@brief The Telemetry message published in this example.
+ * @brief Date-time to use for the model id
  */
-#define sampleazureiotMESSAGE                                 "{\"" sampleazureiotTELEMETRY_NAME "\":%0.2f}"
+#define sampleazureiotDATE_TIME_FORMAT                        "%Y-%m-%dT%H:%M:%S.000Z"
 
+/**
+ * @brief The payload to send to the Device Provisioning Service
+ */
+#define sampleazureiotPROVISIONING_PAYLOAD                    "{\"modelId\":\"" sampleazureiotMODEL_ID "\"}"
 
-/* Device values */
-static double xDeviceCurrentTemperature = sampleazureiotDEFAULT_START_TEMP_CELSIUS;
-static double xDeviceMaximumTemperature = sampleazureiotDEFAULT_START_TEMP_CELSIUS;
-static double xDeviceMinimumTemperature = sampleazureiotDEFAULT_START_TEMP_CELSIUS;
-static double xDeviceTemperatureSummation = sampleazureiotDEFAULT_START_TEMP_CELSIUS;
-static uint32_t ulDeviceTemperatureCount = sampleazureiotDEFAULT_START_TEMP_COUNT;
-static double xDeviceAverageTemperature = sampleazureiotDEFAULT_START_TEMP_CELSIUS;
+/**
+ * @brief Time in ticks to wait between each cycle of the demo implemented
+ * by prvMQTTDemoTask().
+ */
+#define sampleazureiotDELAY_BETWEEN_DEMO_ITERATIONS_TICKS     ( pdMS_TO_TICKS( 5000U ) )
+
+/**
+ * @brief Timeout for MQTT_ProcessLoop in milliseconds.
+ */
+#define sampleazureiotPROCESS_LOOP_TIMEOUT_MS                 ( 500U )
+
+/**
+ * @brief Delay (in ticks) between consecutive cycles of MQTT publish operations in a
+ * demo iteration.
+ *
+ * Note that the process loop also has a timeout, so the total time between
+ * publishes is the sum of the two delays.
+ */
+#define sampleazureiotDELAY_BETWEEN_PUBLISHES_TICKS           ( pdMS_TO_TICKS( 2000U ) )
+
+/**
+ * @brief Transport timeout in milliseconds for transport send and receive.
+ */
+#define sampleazureiotTRANSPORT_SEND_RECV_TIMEOUT_MS          ( 2000U )
+
+/**
+ * @brief Transport timeout in milliseconds for transport send and receive.
+ */
+#define sampleazureiotProvisioning_Registration_TIMEOUT_MS    ( 3 * 1000U )
+
+/**
+ * @brief Wait timeout for subscribe to finish.
+ */
+#define sampleazureiotSUBSCRIBE_TIMEOUT                       ( 10 * 1000U )
+/*-----------------------------------------------------------*/
+
+/**
+ * @brief Unix time.
+ *
+ * @return Time in milliseconds.
+ */
+uint64_t ullGetUnixTime( void );
+/*-----------------------------------------------------------*/
+
+/* Define buffer for IoT Hub info.  */
+#ifdef democonfigENABLE_DPS_SAMPLE
+    static uint8_t ucSampleIotHubHostname[ 128 ];
+    static uint8_t ucSampleIotHubDeviceId[ 128 ];
+    static AzureIoTProvisioningClient_t xAzureIoTProvisioningClient;
+#endif /* democonfigENABLE_DPS_SAMPLE */
+
+/* Each compilation unit must define the NetworkContext struct. */
+struct NetworkContext
+{
+    TlsTransportParams_t * pParams;
+};
+
+AzureIoTHubClient_t xAzureIoTHubClient;
+
+/* Telemetry buffers */
+static uint8_t ucScratchBuffer[ 512 ];
 
 /* Command buffers */
-static uint8_t ucCommandStartTimeValueBuffer[ 32 ];
+static uint8_t ucCommandResponsePayloadBuffer[ 256 ];
 
-/* Property buffers */
-static uint8_t ucPropertyPayloadBuffer[ 256 ];
+/* Reported Properties buffers */
+static uint8_t pucReportedPropertiesUpdate[ 128 ];
+static uint32_t ulReportedPropertiesUpdateLength;
+/*-----------------------------------------------------------*/
 
+#ifdef democonfigENABLE_DPS_SAMPLE
+
+/**
+ * @brief Gets the IoT Hub endpoint and deviceId from Provisioning service.
+ *   This function will block for Provisioning service for result or return failure.
+ *
+ * @param[in] pXNetworkCredentials  Network credential used to connect to Provisioning service
+ * @param[out] ppucIothubHostname  Pointer to uint8_t* IoT Hub hostname return from Provisioning Service
+ * @param[in,out] pulIothubHostnameLength  Length of hostname
+ * @param[out] ppucIothubDeviceId  Pointer to uint8_t* deviceId return from Provisioning Service
+ * @param[in,out] pulIothubDeviceIdLength  Length of deviceId
+ */
+    static uint32_t prvIoTHubInfoGet( NetworkCredentials_t * pXNetworkCredentials,
+                                      uint8_t ** ppucIothubHostname,
+                                      uint32_t * pulIothubHostnameLength,
+                                      uint8_t ** ppucIothubDeviceId,
+                                      uint32_t * pulIothubDeviceIdLength );
+
+#endif /* democonfigENABLE_DPS_SAMPLE */
+
+/**
+ * @brief The task used to demonstrate the Azure IoT Hub API.
+ *
+ * @param[in] pvParameters Parameters as passed at the time of task creation. Not
+ * used in this example.
+ */
+static void prvAzureDemoTask( void * pvParameters );
+
+/**
+ * @brief Connect to endpoint with reconnection retries.
+ *
+ * If connection fails, retry is attempted after a timeout.
+ * Timeout value will exponentially increase until maximum
+ * timeout value is reached or the number of attempts are exhausted.
+ *
+ * @param pcHostName Hostname of the endpoint to connect to.
+ * @param ulPort Endpoint port.
+ * @param pxNetworkCredentials Pointer to Network credentials.
+ * @param pxNetworkContext Point to Network context created.
+ * @return uint32_t The status of the final connection attempt.
+ */
+static uint32_t prvConnectToServerWithBackoffRetries( const char * pcHostName,
+                                                      uint32_t ulPort,
+                                                      NetworkCredentials_t * pxNetworkCredentials,
+                                                      NetworkContext_t * pxNetworkContext );
 /*-----------------------------------------------------------*/
 
 /**
- * @brief Generate max min payload.
+ * @brief Static buffer used to hold MQTT messages being sent and received.
  */
-static AzureIoTResult_t prvInvokeMaxMinCommand( AzureIoTJSONReader_t * pxReader,
-                                                AzureIoTJSONWriter_t * pxWriter )
-{
-    AzureIoTResult_t xResult;
-    uint32_t ulSinceTimeLength;
-
-    /* Get the start time */
-    if( ( xResult = AzureIoTJSONReader_NextToken( pxReader ) )
-        != eAzureIoTSuccess )
-    {
-        LogError( ( "Error getting next token: result 0x%08x", xResult ) );
-    }
-    else if( ( xResult = AzureIoTJSONReader_GetTokenString( pxReader,
-                                                            ucCommandStartTimeValueBuffer,
-                                                            sizeof( ucCommandStartTimeValueBuffer ),
-                                                            &ulSinceTimeLength ) )
-             != eAzureIoTSuccess )
-    {
-        LogError( ( "Error getting token string: result 0x%08x", xResult ) );
-    }
-    else if( ( xResult = AzureIoTJSONWriter_AppendBeginObject( pxWriter ) )
-             != eAzureIoTSuccess )
-    {
-        LogError( ( "Error appending begin object: result 0x%08x", xResult ) );
-    }
-    else if( ( xResult = AzureIoTJSONWriter_AppendPropertyWithDoubleValue( pxWriter, ( const uint8_t * ) sampleazureiotCOMMAND_MAX_TEMP,
-                                                                           sizeof( sampleazureiotCOMMAND_MAX_TEMP ) - 1,
-                                                                           xDeviceMaximumTemperature, sampleazureiotDOUBLE_DECIMAL_PLACE_DIGITS ) )
-             != eAzureIoTSuccess )
-    {
-        LogError( ( "Error appending max temp: result 0x%08x", xResult ) );
-    }
-    else if( ( xResult = AzureIoTJSONWriter_AppendPropertyWithDoubleValue( pxWriter, ( const uint8_t * ) sampleazureiotCOMMAND_MIN_TEMP,
-                                                                           sizeof( sampleazureiotCOMMAND_MIN_TEMP ) - 1,
-                                                                           xDeviceMinimumTemperature, sampleazureiotDOUBLE_DECIMAL_PLACE_DIGITS ) )
-             != eAzureIoTSuccess )
-    {
-        LogError( ( "Error appending min temp: result 0x%08x", xResult ) );
-    }
-    else if( ( xResult = AzureIoTJSONWriter_AppendPropertyWithDoubleValue( pxWriter, ( const uint8_t * ) sampleazureiotCOMMAND_TEMP_VERSION,
-                                                                           sizeof( sampleazureiotCOMMAND_TEMP_VERSION ) - 1,
-                                                                           xDeviceAverageTemperature, sampleazureiotDOUBLE_DECIMAL_PLACE_DIGITS ) )
-             != eAzureIoTSuccess )
-    {
-        LogError( ( "Error appending average temp: result 0x%08x", xResult ) );
-    }
-    else if( ( xResult = AzureIoTJSONWriter_AppendPropertyWithStringValue( pxWriter, ( const uint8_t * ) sampleazureiotCOMMAND_START_TIME,
-                                                                           sizeof( sampleazureiotCOMMAND_START_TIME ) - 1,
-                                                                           ucCommandStartTimeValueBuffer, ulSinceTimeLength ) )
-             != eAzureIoTSuccess )
-    {
-        LogError( ( "Error appending start time: result 0x%08x", xResult ) );
-    }
-    /* Faking the end time to simplify dependencies on <time.h> */
-    else if( ( xResult = AzureIoTJSONWriter_AppendPropertyWithStringValue( pxWriter, ( const uint8_t * ) sampleazureiotCOMMAND_END_TIME,
-                                                                           sizeof( sampleazureiotCOMMAND_END_TIME ) - 1,
-                                                                           ( const uint8_t * ) sampleazureiotCOMMAND_FAKE_END_TIME,
-                                                                           sizeof( sampleazureiotCOMMAND_FAKE_END_TIME ) - 1 ) )
-             != eAzureIoTSuccess )
-    {
-        LogError( ( "Error appending end time: result 0x%08x", xResult ) );
-    }
-    else if( ( xResult = AzureIoTJSONWriter_AppendEndObject( pxWriter ) )
-             != eAzureIoTSuccess )
-    {
-        LogError( ( "Error appending end object: result 0x%08x", xResult ) );
-    }
-
-    return xResult;
-}
-
-/*-----------------------------------------------------------*/
-
-static void prvSkipPropertyAndValue( AzureIoTJSONReader_t * pxReader )
-{
-    AzureIoTResult_t xResult;
-
-    xResult = AzureIoTJSONReader_NextToken( pxReader );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONReader_SkipChildren( pxReader );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONReader_NextToken( pxReader );
-    configASSERT( xResult == eAzureIoTSuccess );
-}
-/*-----------------------------------------------------------*/
+static uint8_t ucMQTTMessageBuffer[ democonfigNETWORK_BUFFER_SIZE ];
 
 /**
- * @brief Update local device temperature values based on new requested temperature.
+ * @brief Internal function for handling Command requests.
+ * 
+ * @remark This function is required for the interface with samples to work properly.
  */
-static void prvUpdateLocalProperties( double xNewTemperatureValue,
-                                      uint32_t ulPropertyVersion,
-                                      bool * pxOutMaxTempChanged )
+static void prvHandleCommand( AzureIoTHubClientCommandRequest_t * pxMessage,
+                              void * pvContext )
 {
-    *pxOutMaxTempChanged = false;
-    xDeviceCurrentTemperature = xNewTemperatureValue;
-
-    /* Update maximum or minimum temperatures. */
-    if( xDeviceCurrentTemperature > xDeviceMaximumTemperature )
-    {
-        xDeviceMaximumTemperature = xDeviceCurrentTemperature;
-        *pxOutMaxTempChanged = true;
-    }
-    else if( xDeviceCurrentTemperature < xDeviceMinimumTemperature )
-    {
-        xDeviceMinimumTemperature = xDeviceCurrentTemperature;
-    }
-
-    /* Calculate the new average temperature. */
-    ulDeviceTemperatureCount++;
-    xDeviceTemperatureSummation += xDeviceCurrentTemperature;
-    xDeviceAverageTemperature = xDeviceTemperatureSummation / ulDeviceTemperatureCount;
-
-    LogInfo( ( "Client updated desired temperature variables locally." ) );
-    LogInfo( ( "Current Temperature: %2f", xDeviceCurrentTemperature ) );
-    LogInfo( ( "Maximum Temperature: %2f", xDeviceMaximumTemperature ) );
-    LogInfo( ( "Minimum Temperature: %2f", xDeviceMinimumTemperature ) );
-    LogInfo( ( "Average Temperature: %2f", xDeviceAverageTemperature ) );
-}
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Gets the reported properties payload with the maximum temperature value.
- */
-static uint32_t prvGetNewMaxTemp( double xUpdatedTemperature, 
-                                  uint8_t * ucReportedPropertyPayloadBuffer,
-                                  uint32_t ulReportedPropertyPayloadBufferSize )
-{
+    AzureIoTHubClient_t * pxHandle = ( AzureIoTHubClient_t * ) pvContext;
+    uint32_t ulResponseStatus = 0;
     AzureIoTResult_t xResult;
-    AzureIoTJSONWriter_t xWriter;
-    int32_t lBytesWritten;
 
-    /* Initialize the JSON writer with the buffer to which we will write the payload with the new temperature. */
-    xResult = AzureIoTJSONWriter_Init( &xWriter, ucReportedPropertyPayloadBuffer, ulReportedPropertyPayloadBufferSize );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONWriter_AppendBeginObject( &xWriter );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONWriter_AppendPropertyName( &xWriter, ( const uint8_t * ) sampleazureiotPROPERTY_MAX_TEMPERATURE_TEXT,
-                                                     sizeof( sampleazureiotPROPERTY_MAX_TEMPERATURE_TEXT ) - 1 );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONWriter_AppendDouble( &xWriter, xUpdatedTemperature, sampleazureiotDOUBLE_DECIMAL_PLACE_DIGITS );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONWriter_AppendEndObject( &xWriter );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    lBytesWritten = AzureIoTJSONWriter_GetBytesUsed( &xWriter );
-
-    return lBytesWritten;
-}
-
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Send updated maximum temperature value to IoT Hub.
- */
-static void prvSendNewMaxTemp( double xUpdatedTemperature )
-{
-    AzureIoTResult_t xResult;
-    int32_t lBytesWritten = prvGetNewMaxTemp( xUpdatedTemperature, ucPropertyPayloadBuffer, sizeof( ucPropertyPayloadBuffer ) );
-
-    if( lBytesWritten < 0 )
+    uint32_t ulCommandResponsePayloadLength = ulHandleCommand( pxMessage, 
+                                               &ulResponseStatus,
+                                               ucCommandResponsePayloadBuffer, 
+                                               sizeof( ucCommandResponsePayloadBuffer ) );
+    
+    if( ( xResult = AzureIoTHubClient_SendCommandResponse( pxHandle, pxMessage, ulResponseStatus,
+                                                ucCommandResponsePayloadBuffer,
+                                                ulCommandResponsePayloadLength ) ) != eAzureIoTSuccess )
     {
-        LogError( ( "Error getting the bytes written for the properties confirmation JSON" ) );
+        LogError( ( "Error sending command response: result 0x%08x", xResult ) );
     }
     else
     {
-        xResult = xUpdateProperties( ucPropertyPayloadBuffer, lBytesWritten );
+        LogInfo( ( "Successfully sent command response %d", ulResponseStatus ) );
+    }
+}
+/*-----------------------------------------------------------*/
 
-        if( xResult != eAzureIoTSuccess )
+/**
+ * @brief Implements the interface for samples to send updates to reported properties.
+ */
+AzureIoTResult_t xUpdateProperties( uint8_t* ucProperties, 
+                                    uint32_t ulPropertiesLength )
+{
+    return AzureIoTHubClient_SendPropertiesReported( &xAzureIoTHubClient, ucProperties, ulPropertiesLength, NULL );
+}
+/*-----------------------------------------------------------*/
+
+/**
+ * @brief Setup transport credentials.
+ */
+static uint32_t prvSetupNetworkCredentials( NetworkCredentials_t * pxNetworkCredentials )
+{
+    pxNetworkCredentials->xDisableSni = pdFALSE;
+    /* Set the credentials for establishing a TLS connection. */
+    pxNetworkCredentials->pucRootCa = ( const unsigned char * ) democonfigROOT_CA_PEM;
+    pxNetworkCredentials->xRootCaSize = sizeof( democonfigROOT_CA_PEM );
+    #ifdef democonfigCLIENT_CERTIFICATE_PEM
+        pxNetworkCredentials->pucClientCert = ( const unsigned char * ) democonfigCLIENT_CERTIFICATE_PEM;
+        pxNetworkCredentials->xClientCertSize = sizeof( democonfigCLIENT_CERTIFICATE_PEM );
+        pxNetworkCredentials->pucPrivateKey = ( const unsigned char * ) democonfigCLIENT_PRIVATE_KEY_PEM;
+        pxNetworkCredentials->xPrivateKeySize = sizeof( democonfigCLIENT_PRIVATE_KEY_PEM );
+    #endif
+
+    return 0;
+}
+/*-----------------------------------------------------------*/
+
+/**
+ * @brief Azure IoT demo task that gets started in the platform specific project.
+ *  In this demo task, middleware API's are used to connect to Azure IoT Hub and
+ *  function to adhere to the Plug and Play device convention.
+ */
+static void prvAzureDemoTask( void * pvParameters )
+{
+    uint32_t ulScratchBufferLength = 0U;
+    NetworkCredentials_t xNetworkCredentials = { 0 };
+    AzureIoTTransportInterface_t xTransport;
+    NetworkContext_t xNetworkContext = { 0 };
+    TlsTransportParams_t xTlsTransportParams = { 0 };
+    AzureIoTResult_t xResult;
+    uint32_t ulStatus;
+    AzureIoTHubClientOptions_t xHubOptions = { 0 };
+    bool xSessionPresent;
+
+    #ifdef democonfigENABLE_DPS_SAMPLE
+        uint8_t * pucIotHubHostname = NULL;
+        uint8_t * pucIotHubDeviceId = NULL;
+        uint32_t pulIothubHostnameLength = 0;
+        uint32_t pulIothubDeviceIdLength = 0;
+    #else
+        uint8_t * pucIotHubHostname = ( uint8_t * ) democonfigHOSTNAME;
+        uint8_t * pucIotHubDeviceId = ( uint8_t * ) democonfigDEVICE_ID;
+        uint32_t pulIothubHostnameLength = sizeof( democonfigHOSTNAME ) - 1;
+        uint32_t pulIothubDeviceIdLength = sizeof( democonfigDEVICE_ID ) - 1;
+    #endif /* democonfigENABLE_DPS_SAMPLE */
+
+    ( void ) pvParameters;
+
+    /* Initialize Azure IoT Middleware.  */
+    configASSERT( AzureIoT_Init() == eAzureIoTSuccess );
+
+    ulStatus = prvSetupNetworkCredentials( &xNetworkCredentials );
+    configASSERT( ulStatus == 0 );
+
+    #ifdef democonfigENABLE_DPS_SAMPLE
+        /* Run DPS.  */
+        if( ( ulStatus = prvIoTHubInfoGet( &xNetworkCredentials, &pucIotHubHostname,
+                                           &pulIothubHostnameLength, &pucIotHubDeviceId,
+                                           &pulIothubDeviceIdLength ) ) != 0 )
         {
-            LogError( ( "There was an error sending the reported properties: result 0x%08x", xResult ) );
+            LogError( ( "Failed on sample_dps_entry!: error code = 0x%08x\r\n", ulStatus ) );
+            return;
         }
-    }
-}
+    #endif /* democonfigENABLE_DPS_SAMPLE */
 
-/*-----------------------------------------------------------*/
+    xNetworkContext.pParams = &xTlsTransportParams;
 
-/**
- * @brief Updates the device's target temperature property online, acknowledging the update from the IoT Hub.
- */
-static void prvAckIncomingTemperature( double xUpdatedTemperature,
-                                       uint32_t ulVersion )
-{
-    AzureIoTResult_t xResult;
-    AzureIoTJSONWriter_t xWriter;
-    int32_t lBytesWritten;
-
-    /* Building the acknowledgement payload for the temperature property to signal we successfully received and accept it. */
-    xResult = AzureIoTJSONWriter_Init( &xWriter, ucPropertyPayloadBuffer, sizeof( ucPropertyPayloadBuffer ) );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONWriter_AppendBeginObject( &xWriter );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTHubClientProperties_BuilderBeginResponseStatus( &xAzureIoTHubClient,
-                                                                      &xWriter,
-                                                                      ( const uint8_t * ) sampleazureiotPROPERTY_TARGET_TEMPERATURE_TEXT,
-                                                                      sizeof( sampleazureiotPROPERTY_TARGET_TEMPERATURE_TEXT ) - 1,
-                                                                      sampleazureiotPROPERTY_STATUS_SUCCESS,
-                                                                      ulVersion,
-                                                                      ( const uint8_t * ) sampleazureiotPROPERTY_SUCCESS,
-                                                                      sizeof( sampleazureiotPROPERTY_SUCCESS ) - 1 );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONWriter_AppendDouble( &xWriter, xUpdatedTemperature, sampleazureiotDOUBLE_DECIMAL_PLACE_DIGITS );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTHubClientProperties_BuilderEndResponseStatus( &xAzureIoTHubClient,
-                                                                    &xWriter );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTJSONWriter_AppendEndObject( &xWriter );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    lBytesWritten = AzureIoTJSONWriter_GetBytesUsed( &xWriter );
-    configASSERT( lBytesWritten > 0 );
-
-    LogDebug( ( "Sending acknowledged writable property. Payload: %.*s", lBytesWritten, ucPropertyPayloadBuffer ) );
-    xResult = xUpdateProperties( ucPropertyPayloadBuffer, lBytesWritten );
-
-    if( xResult != eAzureIoTSuccess )
+    for( ; ; )
     {
-        LogError( ( "There was an error sending the reported properties: 0x%08x", xResult ) );
-    }
-}
+        /* Attempt to establish TLS session with IoT Hub. If connection fails,
+         * retry after a timeout. Timeout value will be exponentially increased
+         * until  the maximum number of attempts are reached or the maximum timeout
+         * value is reached. The function returns a failure status if the TCP
+         * connection cannot be established to the IoT Hub after the configured
+         * number of attempts. */
+        ulStatus = prvConnectToServerWithBackoffRetries( ( const char * ) pucIotHubHostname,
+                                                         democonfigIOTHUB_PORT,
+                                                         &xNetworkCredentials, &xNetworkContext );
+        configASSERT( ulStatus == 0 );
 
-/*-----------------------------------------------------------*/
+        /* Fill in Transport Interface send and receive function pointers. */
+        xTransport.pxNetworkContext = &xNetworkContext;
+        xTransport.xSend = TLS_Socket_Send;
+        xTransport.xRecv = TLS_Socket_Recv;
 
-/**
- * @brief Properties callback handler
- */
-static AzureIoTResult_t prvProcessProperties( AzureIoTHubClientPropertiesResponse_t * pxMessage,
-                                              AzureIoTHubClientPropertyType_t xPropertyType,
-                                              double * pxOutTemperature,
-                                              uint32_t * ulOutVersion )
-{
-    AzureIoTResult_t xResult;
-    AzureIoTJSONReader_t xReader;
-    const uint8_t * pucComponentName = NULL;
-    uint32_t ulComponentNameLength = 0;
-
-    *pxOutTemperature = 0.0;
-
-    xResult = AzureIoTJSONReader_Init( &xReader, pxMessage->pvMessagePayload, pxMessage->ulPayloadLength );
-    configASSERT( xResult == eAzureIoTSuccess );
-
-    xResult = AzureIoTHubClientProperties_GetPropertiesVersion( &xAzureIoTHubClient, &xReader, pxMessage->xMessageType, ulOutVersion );
-
-    if( xResult != eAzureIoTSuccess )
-    {
-        LogError( ( "Error getting the property version: result 0x%08x", xResult ) );
-    }
-    else
-    {
-        /* Reset JSON reader to the beginning */
-        xResult = AzureIoTJSONReader_Init( &xReader, pxMessage->pvMessagePayload, pxMessage->ulPayloadLength );
+        /* Init IoT Hub option */
+        xResult = AzureIoTHubClient_OptionsInit( &xHubOptions );
         configASSERT( xResult == eAzureIoTSuccess );
 
-        while( ( xResult = AzureIoTHubClientProperties_GetNextComponentProperty( &xAzureIoTHubClient, &xReader,
-                                                                                 pxMessage->xMessageType, xPropertyType,
-                                                                                 &pucComponentName, &ulComponentNameLength ) ) == eAzureIoTSuccess )
+        xHubOptions.pucModuleID = ( const uint8_t * ) democonfigMODULE_ID;
+        xHubOptions.ulModuleIDLength = sizeof( democonfigMODULE_ID ) - 1;
+        xHubOptions.pucModelID = ( const uint8_t * ) sampleazureiotMODEL_ID;
+        xHubOptions.ulModelIDLength = sizeof( sampleazureiotMODEL_ID ) - 1;
+
+        xResult = AzureIoTHubClient_Init( &xAzureIoTHubClient,
+                                          pucIotHubHostname, pulIothubHostnameLength,
+                                          pucIotHubDeviceId, pulIothubDeviceIdLength,
+                                          &xHubOptions,
+                                          ucMQTTMessageBuffer, sizeof( ucMQTTMessageBuffer ),
+                                          ullGetUnixTime,
+                                          &xTransport );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        #ifdef democonfigDEVICE_SYMMETRIC_KEY
+            xResult = AzureIoTHubClient_SetSymmetricKey( &xAzureIoTHubClient,
+                                                         ( const uint8_t * ) democonfigDEVICE_SYMMETRIC_KEY,
+                                                         sizeof( democonfigDEVICE_SYMMETRIC_KEY ) - 1,
+                                                         Crypto_HMAC );
+            configASSERT( xResult == eAzureIoTSuccess );
+        #endif /* democonfigDEVICE_SYMMETRIC_KEY */
+
+        /* Sends an MQTT Connect packet over the already established TLS connection,
+         * and waits for connection acknowledgment (CONNACK) packet. */
+        LogInfo( ( "Creating an MQTT connection to %s.\r\n", pucIotHubHostname ) );
+
+        xResult = AzureIoTHubClient_Connect( &xAzureIoTHubClient,
+                                             false, &xSessionPresent,
+                                             sampleazureiotCONNACK_RECV_TIMEOUT_MS );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        xResult = AzureIoTHubClient_SubscribeCommand( &xAzureIoTHubClient, prvHandleCommand,
+                                                      &xAzureIoTHubClient, sampleazureiotSUBSCRIBE_TIMEOUT );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        xResult = AzureIoTHubClient_SubscribeProperties( &xAzureIoTHubClient, vHandleProperties,
+                                                         &xAzureIoTHubClient, sampleazureiotSUBSCRIBE_TIMEOUT );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        /* Get property document after initial connection */
+        xResult = AzureIoTHubClient_RequestPropertiesAsync( &xAzureIoTHubClient );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        /* Publish messages with QoS1, send and process Keep alive messages. */
+        for( ; ; )
         {
-            if( ulComponentNameLength > 0 )
+            /* Hook for sending Telemetry */
+            if (ulCreateTelemetry(ucScratchBuffer, sizeof( ucScratchBuffer ), &ulScratchBufferLength ) == 0 && 
+                ulScratchBufferLength > 0)
             {
-                LogInfo( ( "Unknown component name received" ) );
-
-                /* Unknown component name arrived (there are none for this device).
-                 * We have to skip over the property and value to continue iterating */
-                prvSkipPropertyAndValue( &xReader );
-            }
-            else if( AzureIoTJSONReader_TokenIsTextEqual( &xReader,
-                                                          ( const uint8_t * ) sampleazureiotPROPERTY_TARGET_TEMPERATURE_TEXT,
-                                                          sizeof( sampleazureiotPROPERTY_TARGET_TEMPERATURE_TEXT ) - 1 ) )
-            {
-                xResult = AzureIoTJSONReader_NextToken( &xReader );
-                configASSERT( xResult == eAzureIoTSuccess );
-
-                /* Get desired temperature */
-                xResult = AzureIoTJSONReader_GetTokenDouble( &xReader, pxOutTemperature );
-
-                if( xResult != eAzureIoTSuccess )
-                {
-                    LogError( ( "Error getting the property version: result 0x%08x", xResult ) );
-                    break;
-                }
-
-                xResult = AzureIoTJSONReader_NextToken( &xReader );
+                xResult = AzureIoTHubClient_SendTelemetry( &xAzureIoTHubClient,
+                                                        ucScratchBuffer, ulScratchBufferLength,
+                                                        NULL, eAzureIoTHubMessageQoS1, NULL );
                 configASSERT( xResult == eAzureIoTSuccess );
             }
-            else
+
+            /* Hook for sending update to reported properties */
+            ulReportedPropertiesUpdateLength = ulCreateReportedPropertiesUpdate( pucReportedPropertiesUpdate, sizeof( pucReportedPropertiesUpdate ) );
+
+            if (ulReportedPropertiesUpdateLength > 0)
             {
-                LogInfo( ( "Unknown property arrived: skipping over it." ) );
-
-                /* Unknown property arrived. We have to skip over the property and value to continue iterating. */
-                prvSkipPropertyAndValue( &xReader );
+                xResult = xUpdateProperties(pucReportedPropertiesUpdate, ulReportedPropertiesUpdateLength);
+            
+                configASSERT( xResult == eAzureIoTSuccess );
             }
+
+            LogInfo( ( "Attempt to receive publish message from IoT Hub.\r\n" ) );
+            xResult = AzureIoTHubClient_ProcessLoop( &xAzureIoTHubClient,
+                                                     sampleazureiotPROCESS_LOOP_TIMEOUT_MS );
+            configASSERT( xResult == eAzureIoTSuccess );
+
+            /* Leave Connection Idle for some time. */
+            LogInfo( ( "Keeping Connection Idle...\r\n\r\n" ) );
+            vTaskDelay( sampleazureiotDELAY_BETWEEN_PUBLISHES_TICKS );
         }
 
-        if( xResult != eAzureIoTErrorEndOfProperties )
-        {
-            LogError( ( "There was an error parsing the properties: result 0x%08x", xResult ) );
-        }
-        else
-        {
-            LogInfo( ( "Successfully parsed properties" ) );
-            xResult = eAzureIoTSuccess;
-        }
-    }
-
-    return xResult;
-}
-
-/*-----------------------------------------------------------*/
-
-static void prvHandlePropertyUpdate( AzureIoTHubClientPropertiesResponse_t * pxMessage )
-{
-    AzureIoTResult_t xResult;
-    double xIncomingTemperature;
-    uint32_t ulVersion;
-    bool xWasMaxTemperatureChanged = false;
-
-    xResult = prvProcessProperties( pxMessage, eAzureIoTHubClientPropertyWritable, &xIncomingTemperature, &ulVersion );
-
-    if( xResult == eAzureIoTSuccess )
-    {
-        prvUpdateLocalProperties( xIncomingTemperature, ulVersion, &xWasMaxTemperatureChanged );
-        prvAckIncomingTemperature( xIncomingTemperature, ulVersion );
-
-        if( xWasMaxTemperatureChanged )
-        {
-            prvSendNewMaxTemp( xIncomingTemperature );
-        }
-    }
-    else
-    {
-        LogError( ( "There was an error processing incoming properties: result 0x%08x", xResult ) );
-    }
-}
-
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Property message callback handler
- */
-void vHandleProperties( AzureIoTHubClientPropertiesResponse_t * pxMessage,
-                        void * pvContext )
-{
-    ( void ) pvContext;
-
-    LogDebug( ( "Property document payload : %.*s \r\n",
-                pxMessage->ulPayloadLength,
-                ( const char * ) pxMessage->pvMessagePayload ) );
-
-    switch( pxMessage->xMessageType )
-    {
-        case eAzureIoTHubPropertiesRequestedMessage:
-            LogDebug( ( "Device property document GET received" ) );
-
-            prvHandlePropertyUpdate( pxMessage );
-
-            break;
-
-        case eAzureIoTHubPropertiesWritablePropertyMessage:
-            LogDebug( ( "Device writeable property received" ) );
-
-            prvHandlePropertyUpdate( pxMessage );
-
-            break;
-
-        case eAzureIoTHubPropertiesReportedResponseMessage:
-            LogDebug( ( "Device reported property response received" ) );
-            break;
-
-        default:
-            LogError( ( "Unknown property message: 0x%08x", pxMessage->xMessageType ) );
-            configASSERT( false );
-    }
-}
-
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Command message callback handler
- */
-uint32_t ulHandleCommand( AzureIoTHubClientCommandRequest_t * pxMessage,
-                          uint32_t* pulResponseStatus,
-                          uint8_t* pucCommandResponsePayloadBuffer,
-                          uint32_t ulCommandResponsePayloadBufferSize)
-{
-    AzureIoTResult_t xResult;
-    AzureIoTJSONReader_t xReader;
-    AzureIoTJSONWriter_t xWriter;
-    int32_t lCommandNameLength;
-    int32_t ulCommandResponsePayloadLength;
-
-    LogInfo( ( "Command payload : %.*s \r\n",
-               pxMessage->ulPayloadLength,
-               ( const char * ) pxMessage->pvMessagePayload ) );
-
-    lCommandNameLength = sizeof( sampleazureiotCOMMAND_MAX_MIN_REPORT ) - 1;
-
-    if( ( lCommandNameLength == pxMessage->usCommandNameLength ) &&
-        ( strncmp( sampleazureiotCOMMAND_MAX_MIN_REPORT, ( const char * ) pxMessage->pucCommandName, lCommandNameLength ) == 0 ) )
-    {
-        /* Is for max min report */
-
-        /*Initialize the reader from which we pull the "since". */
-        xResult = AzureIoTJSONReader_Init( &xReader, pxMessage->pvMessagePayload, pxMessage->ulPayloadLength );
+        xResult = AzureIoTHubClient_UnsubscribeProperties( &xAzureIoTHubClient );
         configASSERT( xResult == eAzureIoTSuccess );
 
-        /* Initialize the JSON writer with a buffer to which we will write the response payload. */
-        xResult = AzureIoTJSONWriter_Init( &xWriter, pucCommandResponsePayloadBuffer, ulCommandResponsePayloadBufferSize );
+        xResult = AzureIoTHubClient_UnsubscribeCommand( &xAzureIoTHubClient );
         configASSERT( xResult == eAzureIoTSuccess );
 
-        /* Read from the writer the "since" value and use it to construct the response payload in the writer. */
-        xResult = prvInvokeMaxMinCommand( &xReader, &xWriter );
+        /* Send an MQTT Disconnect packet over the already connected TLS over
+         * TCP connection. There is no corresponding response for the disconnect
+         * packet. After sending disconnect, client must close the network
+         * connection. */
+        xResult = AzureIoTHubClient_Disconnect( &xAzureIoTHubClient );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        /* Close the network connection.  */
+        TLS_Socket_Disconnect( &xNetworkContext );
+
+        /* Wait for some time between two iterations to ensure that we do not
+         * bombard the IoT Hub. */
+        LogInfo( ( "Demo completed successfully.\r\n" ) );
+        LogInfo( ( "Short delay before starting the next iteration.... \r\n\r\n" ) );
+        vTaskDelay( sampleazureiotDELAY_BETWEEN_DEMO_ITERATIONS_TICKS );
+    }
+}
+/*-----------------------------------------------------------*/
+
+#ifdef democonfigENABLE_DPS_SAMPLE
+
+/**
+ * @brief Get IoT Hub endpoint and device Id info, when Provisioning service is used.
+ *   This function will block for Provisioning service for result or return failure.
+ */
+    static uint32_t prvIoTHubInfoGet( NetworkCredentials_t * pXNetworkCredentials,
+                                      uint8_t ** ppucIothubHostname,
+                                      uint32_t * pulIothubHostnameLength,
+                                      uint8_t ** ppucIothubDeviceId,
+                                      uint32_t * pulIothubDeviceIdLength )
+    {
+        NetworkContext_t xNetworkContext = { 0 };
+        TlsTransportParams_t xTlsTransportParams = { 0 };
+        AzureIoTResult_t xResult;
+        AzureIoTTransportInterface_t xTransport;
+        uint32_t ucSamplepIothubHostnameLength = sizeof( ucSampleIotHubHostname );
+        uint32_t ucSamplepIothubDeviceIdLength = sizeof( ucSampleIotHubDeviceId );
+        uint32_t ulStatus;
+
+        /* Set the pParams member of the network context with desired transport. */
+        xNetworkContext.pParams = &xTlsTransportParams;
+
+        ulStatus = prvConnectToServerWithBackoffRetries( democonfigENDPOINT, democonfigIOTHUB_PORT,
+                                                         pXNetworkCredentials, &xNetworkContext );
+        configASSERT( ulStatus == 0 );
+
+        /* Fill in Transport Interface send and receive function pointers. */
+        xTransport.pxNetworkContext = &xNetworkContext;
+        xTransport.xSend = TLS_Socket_Send;
+        xTransport.xRecv = TLS_Socket_Recv;
+
+        xResult = AzureIoTProvisioningClient_Init( &xAzureIoTProvisioningClient,
+                                                   ( const uint8_t * ) democonfigENDPOINT,
+                                                   sizeof( democonfigENDPOINT ) - 1,
+                                                   ( const uint8_t * ) democonfigID_SCOPE,
+                                                   sizeof( democonfigID_SCOPE ) - 1,
+                                                   ( const uint8_t * ) democonfigREGISTRATION_ID,
+                                                   sizeof( democonfigREGISTRATION_ID ) - 1,
+                                                   NULL, ucMQTTMessageBuffer, sizeof( ucMQTTMessageBuffer ),
+                                                   ullGetUnixTime,
+                                                   &xTransport );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        #ifdef democonfigDEVICE_SYMMETRIC_KEY
+            xResult = AzureIoTProvisioningClient_SetSymmetricKey( &xAzureIoTProvisioningClient,
+                                                                  ( const uint8_t * ) democonfigDEVICE_SYMMETRIC_KEY,
+                                                                  sizeof( democonfigDEVICE_SYMMETRIC_KEY ) - 1,
+                                                                  Crypto_HMAC );
+            configASSERT( xResult == eAzureIoTSuccess );
+        #endif /* democonfigDEVICE_SYMMETRIC_KEY */
+
+        xResult = AzureIoTProvisioningClient_SetRegistrationPayload( &xAzureIoTProvisioningClient,
+                                                                     ( const uint8_t * ) sampleazureiotPROVISIONING_PAYLOAD,
+                                                                     sizeof( sampleazureiotPROVISIONING_PAYLOAD ) - 1 );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        do
+        {
+            xResult = AzureIoTProvisioningClient_Register( &xAzureIoTProvisioningClient,
+                                                           sampleazureiotProvisioning_Registration_TIMEOUT_MS );
+        } while( xResult == eAzureIoTErrorPending );
 
         if( xResult == eAzureIoTSuccess )
         {
-            ulCommandResponsePayloadLength = AzureIoTJSONWriter_GetBytesUsed( &xWriter );
-
-            *pulResponseStatus = 200;
+            LogInfo( ( "Successfully acquired IoT Hub name and Device ID" ) );
         }
         else
         {
-            LogError( ( "Error generating command payload: result 0x%08x", xResult ) );
-
-            *pulResponseStatus = 501;
-            ulCommandResponsePayloadLength = sizeof( sampleazureiotCOMMAND_EMPTY_PAYLOAD ) - 1;
-            (void)memcpy( pucCommandResponsePayloadBuffer, sampleazureiotCOMMAND_EMPTY_PAYLOAD, ulCommandResponsePayloadLength );
+            LogInfo( ( "Error geting IoT Hub name and Device ID: 0x%08x", xResult ) );
         }
+
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        xResult = AzureIoTProvisioningClient_GetDeviceAndHub( &xAzureIoTProvisioningClient,
+                                                              ucSampleIotHubHostname, &ucSamplepIothubHostnameLength,
+                                                              ucSampleIotHubDeviceId, &ucSamplepIothubDeviceIdLength );
+        configASSERT( xResult == eAzureIoTSuccess );
+
+        AzureIoTProvisioningClient_Deinit( &xAzureIoTProvisioningClient );
+
+        /* Close the network connection.  */
+        TLS_Socket_Disconnect( &xNetworkContext );
+
+        *ppucIothubHostname = ucSampleIotHubHostname;
+        *pulIothubHostnameLength = ucSamplepIothubHostnameLength;
+        *ppucIothubDeviceId = ucSampleIotHubDeviceId;
+        *pulIothubDeviceIdLength = ucSamplepIothubDeviceIdLength;
+
+        return 0;
     }
-    else
+
+#endif /* democonfigENABLE_DPS_SAMPLE */
+/*-----------------------------------------------------------*/
+
+/**
+ * @brief Connect to server with backoff retries.
+ */
+static uint32_t prvConnectToServerWithBackoffRetries( const char * pcHostName,
+                                                      uint32_t port,
+                                                      NetworkCredentials_t * pxNetworkCredentials,
+                                                      NetworkContext_t * pxNetworkContext )
+{
+    TlsTransportStatus_t xNetworkStatus;
+    BackoffAlgorithmStatus_t xBackoffAlgStatus = BackoffAlgorithmSuccess;
+    BackoffAlgorithmContext_t xReconnectParams;
+    uint16_t usNextRetryBackOff = 0U;
+
+    /* Initialize reconnect attempts and interval. */
+    BackoffAlgorithm_InitializeParams( &xReconnectParams,
+                                       sampleazureiotRETRY_BACKOFF_BASE_MS,
+                                       sampleazureiotRETRY_MAX_BACKOFF_DELAY_MS,
+                                       sampleazureiotRETRY_MAX_ATTEMPTS );
+
+    /* Attempt to connect to IoT Hub. If connection fails, retry after
+     * a timeout. Timeout value will exponentially increase till maximum
+     * attempts are reached.
+     */
+    do
     {
-        /* Not for max min report (not for this device) */
-        LogInfo( ( "Received command is not for this device: %.*s",
-                   pxMessage->usCommandNameLength,
-                   pxMessage->pucCommandName ) );
+        LogInfo( ( "Creating a TLS connection to %s:%u.\r\n", pcHostName, port ) );
+        /* Attempt to create a mutually authenticated TLS connection. */
+        xNetworkStatus = TLS_Socket_Connect( pxNetworkContext,
+                                             pcHostName, port,
+                                             pxNetworkCredentials,
+                                             sampleazureiotTRANSPORT_SEND_RECV_TIMEOUT_MS,
+                                             sampleazureiotTRANSPORT_SEND_RECV_TIMEOUT_MS );
 
-        *pulResponseStatus = 404;
-        ulCommandResponsePayloadLength = sizeof( sampleazureiotCOMMAND_EMPTY_PAYLOAD ) - 1;
-        (void)memcpy( pucCommandResponsePayloadBuffer, sampleazureiotCOMMAND_EMPTY_PAYLOAD, ulCommandResponsePayloadLength );
-    }
+        if( xNetworkStatus != eTLSTransportSuccess )
+        {
+            /* Generate a random number and calculate backoff value (in milliseconds) for
+             * the next connection retry.
+             * Note: It is recommended to seed the random number generator with a device-specific
+             * entropy source so that possibility of multiple devices retrying failed network operations
+             * at similar intervals can be avoided. */
+            xBackoffAlgStatus = BackoffAlgorithm_GetNextBackoff( &xReconnectParams, configRAND32(), &usNextRetryBackOff );
 
-    return ulCommandResponsePayloadLength;
+            if( xBackoffAlgStatus == BackoffAlgorithmRetriesExhausted )
+            {
+                LogError( ( "Connection to the IoT Hub failed, all attempts exhausted." ) );
+            }
+            else if( xBackoffAlgStatus == BackoffAlgorithmSuccess )
+            {
+                LogWarn( ( "Connection to the IoT Hub failed [%d]. "
+                           "Retrying connection with backoff and jitter [%d]ms.",
+                           xNetworkStatus, usNextRetryBackOff ) );
+                vTaskDelay( pdMS_TO_TICKS( usNextRetryBackOff ) );
+            }
+        }
+    } while( ( xNetworkStatus != eTLSTransportSuccess ) && ( xBackoffAlgStatus == BackoffAlgorithmSuccess ) );
+
+    return xNetworkStatus == eTLSTransportSuccess ? 0 : 1;
 }
-
 /*-----------------------------------------------------------*/
 
-/**
- * @brief Implements the sample interface for generating Telemetry payload.
+/*
+ * @brief Create the task that demonstrates the AzureIoTHub demo
  */
-uint32_t ulCreateTelemetry( uint8_t * pucTelemetryData,
-                            uint32_t ulTelemetryDataLength )
+void vStartDemoTask( void )
 {
-    return snprintf( ( char * ) pucTelemetryData, ulTelemetryDataLength,
-                     sampleazureiotMESSAGE, xDeviceCurrentTemperature );
+    /* This example uses a single application task, which in turn is used to
+     * connect, subscribe, publish, unsubscribe and disconnect from the IoT Hub */
+    xTaskCreate( prvAzureDemoTask,         /* Function that implements the task. */
+                 "AzureDemoTask",          /* Text name for the task - only used for debugging. */
+                 democonfigDEMO_STACKSIZE, /* Size of stack (in words, not bytes) to allocate for the task. */
+                 NULL,                     /* Task parameter - not used in this case. */
+                 tskIDLE_PRIORITY,         /* Task priority, must be between 0 and configMAX_PRIORITIES - 1. */
+                 NULL );                   /* Used to pass out a handle to the created task - not used in this case. */
 }
-
-/*-----------------------------------------------------------*/
-
-/**
- * @brief Implements the sample interface for generating reported properties payload.
- */
-uint32_t ulCreateReportedPropertiesUpdate( uint8_t * pucPropertiesData,
-                                           uint32_t ulPropertiesDataSize )
-{
-    return prvGetNewMaxTemp( xDeviceCurrentTemperature, pucPropertiesData, ulPropertiesDataSize );
-}
-
 /*-----------------------------------------------------------*/
