+++ demos/sample_azure_iot_pnp/sample_azure_iot_pnp.h
@@ -36,13 +36,15 @@ extern AzureIoTHubClient_t xAzureIoTHubClient;
  *         `ulCreateTelemetry` is called periodically by the sample core task (the task created by `vStartDemoTask`). 
  *         If the return is zero, telemetry is not sent to the Azure IoT Hub. 
  * 
- * @param[out] pucTelemetryData    Pointer to uint8_t* that will contain the Telemetry payload.
- * @param[in]  ulTelemetryDataSize Size of `pucTelemetryData`
+ * @param[out] pucTelemetryData        Pointer to uint8_t* that will contain the Telemetry payload.
+ * @param[in]  ulTelemetryDataSize     Size of `pucTelemetryData`
+ * @param[out]  pulTelemetryDataLength The number of bytes written in `pucTelemetryData`
  * 
- * @return uint32_t The number of bytes written in `pucTelemetryData`.
+ * @return uint32_t Zero if successful, non-zero if any failure occurs.
  */
 uint32_t ulCreateTelemetry( uint8_t * pucTelemetryData,
-                            uint32_t ulTelemetryDataSize );
+                            uint32_t ulTelemetryDataSize,
+                            uint32_t * pulTelemetryDataLength );
 
 /**
  * @brief Provides the payload to be sent as reported properties update to the Azure IoT Hub.
